Wrap only `bun build` processes (#21)
`bun build` will keep running even after the parent Elixir process is
closed. There is a wrapper script that monitors stdin and ensures that
bun processes are closed along with the parent Elixir process.

Other bun commands don't leave this empty processes so they can be run
directly without using the wrapper. For example `bun run` executes the
script command in a subshell, ensuring that it is closed appropriately.